pyproject.toml
[tool.poetry]
name = "cloud"
version = "0.1.0"
description = ""
authors = ["Alexey Baranov <atlantij@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
watchdog = {version = "==2.1.3", extras = ["watchmedo"]}
aiotask-context = "==0.6.1"
celery = "==5.1.2"
colorlog = "==5.0.1"
redis = "==3.5.3"
psycopg2-binary = "==2.9.1"
python-dateutil = "==2.8.1"
sorl-thumbnail = "==12.7.0"
uvloop = "==0.15.2"
ujson = "==4.0.2"
uritemplate = "==3.0.1"
amqp = "==5.0.6"
argh = "==0.26.2"
asgiref = "==3.4.1"
billiard = "==3.6.4.0"
certifi = "==2021.5.30"
chardet = "==4.0.0"
click-didyoumean = "==0.0.3"
click-plugins = "==1.1.1"
click-repl = "==0.2.0"
click = "==7.1.2"
coreapi = "==2.3.3"
coreschema = "==0.0.4"
funcy = "==1.16"
future = "==0.18.2"
idna = "==2.10"
inflection = "==0.5.1"
itypes = "==1.2.0"
kombu = "==5.1.0"
prompt-toolkit = "==3.0.19"
pyparsing = "==2.4.7"
python-json-logger = "==2.0.1"
pytz = "==2021.1"
requests = "==2.25.1"
"ruamel.yaml.clib" = "==0.2.6"
"ruamel.yaml" = "==0.17.10"
six = "==1.16.0"
sqlparse = "==0.4.1"
urllib3 = "==1.26.6"
vine = "==5.0.0"
wcwidth = "==0.2.5"
Markdown = "==3.3.4"
Pillow = "==8.3.1"
Pygments = "==2.9.0"
PyJWT = "==2.1.0"
PyYAML = "==5.4.1"
uWSGI = "==2.0.19.1"
Jinja2 = "==3.0.1"
MarkupSafe = "==2.0.1"
django = "==3.2.5"
django-appconf = "==1.0.4"
django-admin-rangefilter = "==0.8.1"
django-cacheops = "==6.0"
django-debug-toolbar = "==3.2.1"
django-extensions = "==3.1.3"
django-filter = "==2.4.0"
django-json-widget = "==1.1.1"
django-redis = "==5.0.0"
django-redis-sessions = "==0.6.2"
django-select2 = "==7.7.1"
django-split-settings = "==1.0.1"
django-celery-email = "==3.0.0"
django-eml-email-backend = "==0.1"
dj-database-url = "==0.5.0"
django-pickling = "==1.0"
django-cors-headers = "==3.7.0"
djangorestframework = "==3.12.4"
drf-yasg = "==1.20.0"
drf-ujson2 = "==1.6.1"
django-mptt = "^0.12.0"
djangorestframework-recursive = "^0.1.2"
django-guardian = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
